+++ src/utils/umami.ts
@@ -5,6 +5,8 @@
  */
 export function trackUmamiEvent(eventName: string, eventProperties?: object) {
   if (typeof window !== "undefined" && "umami" in window) {
-    (window as any).umami.track(eventName, eventProperties);
+    if (eventProperties && Object.keys(eventProperties).length > 0)
+      (window as any).umami.track(eventName, eventProperties);
+    else (window as any).umami.track(eventName);
   }
 }